# services/staffing-service/openapi.yaml
openapi: 3.1.0
info:
  title: CareCareer Staffing Service
  version: 0.1.0
  description: |
    Manages facilities, departments, credential requirements, and workers.
    All endpoints require Bearer token authentication (RS256).
    All operations are tenant-scoped via validated session principal.
servers:
  - url: http://localhost:3200
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/shifts:
    post:
      operationId: createShift
      summary: Create a new shift in DRAFT status
      description: Requires permission shifts:create
      responses:
        '201':
          description: Shift created
    get:
      operationId: listShifts
      summary: List shifts (filterable by status, facility, role)
      description: Requires permission shifts:read
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: facilityId
          in: query
          schema: { type: string, format: uuid }
        - name: role
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Shift list
  /v1/shifts/{shiftId}:
    get:
      operationId: getShift
      summary: Get shift detail
      parameters:
        - name: shiftId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Shift detail
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/shifts/{shiftId}/publish:
    post:
      operationId: publishShift
      summary: Publish a draft shift to marketplace
      description: DRAFT → PUBLISHED. Requires expectedVersion.
      parameters:
        - name: shiftId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Shift published
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/shifts/{shiftId}/cancel:
    post:
      operationId: cancelShift
      summary: Cancel a shift
      description: Any pre-completion state → CANCELLED. Requires reason and expectedVersion.
      parameters:
        - name: shiftId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Shift cancelled
        '409':
          $ref: '#/components/responses/Conflict'

  # ─── GP-09: Marketplace ──────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed platform access token
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
...
  responses:
...
    NotFound:
      description: Resource not found (or hidden cross-tenant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Optimistic concurrency conflict (version mismatch)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'